Home

JuliaForBioinformatics.jl

JuliaForBioinformatics.show_type_tree

JuliaForBioinformatics.show_type_tree — Function.

Recursive function to print the type tree for a given type, including its supertype and all its subtypes.

source

Installation

For the first and second day of the workshop you need to install:

Julia 1.1

Juno, the atom based IDE.

This package with the IJulia notebooks we are going to use:

using Pkg
pkg"add https://github.com/diegozea/JuliaForBioinformatics.git"

IJulia:

using Pkg
pkg"add IJulia"