API

JuliaForBioinformatics' API

JuliaForBioinformatics.show_type_tree

JuliaForBioinformatics.show_type_tree — Function.

Recursive function to print the type tree for a given type, including its supertype and all its subtypes.

source